requirements.txt
# List of required packages (Dependencies)
requests
re
python-telegram-bot 
